+++ packages/bridge-controller/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `UnifiedSwapBridgeEventName.ResponseValidationFailure` event ([#6362](https://github.com/MetaMask/core/pull/6362))
+
 ## [41.1.0]
 
 ### Added